vectordb/pineconedb/example-001/pyproject.toml
[project]
name = "example-001"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.101.0",
    "pandas>=2.3.2",
    "pinecone>=7.3.0",
    "python-dotenv>=1.1.1",
]
